+++ osquery/core/watcher.h
@@ -60,14 +60,14 @@ struct PerformanceState {
   /// A counter of how many intervals the process exceeded performance limits.
   size_t sustained_latency;
   /// The last checked user CPU time.
-  size_t user_time;
+  uint64_t user_time;
   /// The last checked system CPU time.
-  size_t system_time;
+  uint64_t system_time;
   /// A timestamp when the process/worker was last created.
-  size_t last_respawn_time;
+  uint64_t last_respawn_time;
 
   /// The initial (or as close as possible) process image footprint.
-  size_t initial_footprint;
+  uint64_t initial_footprint;
 
   PerformanceState() {
     sustained_latency = 0;
@@ -150,11 +150,11 @@ class Watcher : private boost::noncopyable {
   }
 
   /// Reset counters after a worker exits.
-  void resetWorkerCounters(size_t respawn_time);
+  void resetWorkerCounters(uint64_t respawn_time);
 
   /// Reset counters for an extension path.
   void resetExtensionCounters(const std::string& extension,
-                              size_t respawn_time);
+                              uint64_t respawn_time);
 
   /// Accessor for autoloadable extension paths.
   const ExtensionMap& extensions() const {
@@ -210,11 +210,11 @@ class Watcher : private boost::noncopyable {
     worker_restarts_++;
   }
 
-  void workerStartTime(size_t start_time) {
+  void workerStartTime(uint64_t start_time) {
     worker_start_time_ = start_time;
   }
 
-  size_t workerStartTime() {
+  uint64_t workerStartTime() {
     return worker_start_time_;
   }
 
@@ -230,7 +230,7 @@ class Watcher : private boost::noncopyable {
   std::shared_ptr<PlatformProcess> worker_;
 
   /// Time the worker was started.
-  size_t worker_start_time_{0};
+  uint64_t worker_start_time_{0};
 
   /// Number of worker restarts NOT induced by a watchdog process.
   size_t worker_restarts_{0};
@@ -319,7 +319,7 @@ class WatcherRunner : public InternalRunnable {
   virtual void stopChild(const PlatformProcess& child) const;
 
   /// Return the time the watchdog is delayed until (from start of watcher).
-  size_t delayedTime() const;
+  uint64_t delayedTime() const;
 
  private:
   /// For testing only, ask the WatcherRunner to run a start loop once.
@@ -368,5 +368,5 @@ class WatcherWatcherRunner : public InternalRunnable {
 };
 
 /// Get a performance limit by name and optional level.
-size_t getWorkerLimit(WatchdogLimitType limit);
+uint64_t getWorkerLimit(WatchdogLimitType limit);
 }
